+++ twitchdl/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.11.0"
+__version__ = "1.12.0"
 
 CLIENT_ID = "kimne78kx3ncx6brgo4mv6wki5h1ko"