+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Twitch Downloader change log
 -------------------
 
 * Added `source` as alias for best available quality (#33)
+* Added `--no-join` option to `download` to skip ffmpeg join (#36)
 
 1.11.0 (2020-09-03)
 -------------------